requirements.txt
requests
psycopg2
python-dotenv
retry
streamlit